requirements.txt
dj_database_url==0.5.0
Django==3.2.9
django_embed_video==1.4.0
django_heroku==0.3.1
django_phone_field==1.8.1
python-dotenv==0.19.2
django_filter==2.4.0
whitenoise==5.2.0
gunicorn==20.1.0
requests==2.22.0